# gen/api/openapi.yaml
openapi: 3.0.0
info:
  description: The library service
  title: Library Service
  version: 0.0.1
servers:
- url: /
...
paths:
...
  /payment/updateOrder:
    post:
      description: Refer to summary
      operationId: updateOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/payment_UpdateOrderRequest'
        description: The input body required by this request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payment_UpdateOrderResponse'
          description: Returns updated order upon success
      summary: Endpoint for the Update Order use case
      tags:
      - Payment
      x-contentType: application/json
      x-accepts: application/json
  /payment/submitOrder:
    post:
      description: Refer to summary
      operationId: submitOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/payment_SubmitOrderRequest'
        description: The input body required by this request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payment_SubmitOrderResponse'
          description: Returns submitted order upon success
      summary: Endpoint for the Submit Order use case
      tags:
      - Payment
      x-contentType: application/json
      x-accepts: application/json
...
  /payment/getStatus:
    post:
      description: Refer to summary
      operationId: getStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/payment_GetStatusRequest'
        description: The input body required by this request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payment_GetStatusResponse'
          description: Returns status upon success
      summary: Endpoint for Get Status use case
      tags:
      - Payment
      x-contentType: application/json
      x-accepts: application/json
  /payment/getItems:
    post:
      description: Refer to summary
      operationId: getItemsPayments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/payment_GetItemsRequest'
        description: The input body required by this request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payment_GetItemsResponse'
          description: Returns status upon success
      summary: Endpoint for Get Items use case
      tags:
      - Payment
      x-contentType: application/json
      x-accepts: application/json
...
components:
  schemas:
    ItemObject:
      description: Generic schema for an item
      example:
        itemType: itemType
        quantity: 6
        productId: productId
        size: size
        price: 0.8008281904610115
        imageUrl: imageUrl
        name: name
        description: description
        storeId: storeId
        barcode: barcode
        brand: brand
      properties:
        productId:
          type: string
        name:
          type: string
        barcode:
          type: string
        storeId:
          type: string
        price:
          type: number
        quantity:
          type: integer
        description:
          type: string
        imageUrl:
          type: string
        brand:
          type: string
        size:
          type: string
        itemType:
          type: string
      type: object
...
    payment_UpdateOrderRequest:
      description: This object is expected as input
      properties:
        orderId:
          description: the id of the order to be generated
          type: string
        userId:
          description: generated token used to identify the caller of the endpoint
          type: string
        items:
          items:
            $ref: '#/components/schemas/ItemObject'
          type: array
        discount:
          type: number
        orderType:
          type: string
        deliveryAddress:
          type: string
      type: object
    payment_UpdateOrderResponse:
      description: This object is expected as output
      example:
        success: true
        message: message
        timestamp: timestamp
        order: '{}'
      properties:
        message:
          description: Will contain a descriptive error message if response code is
            not 200, else will be empty
          type: string
        success:
          type: boolean
        timestamp:
          type: string
        order:
          type: object
      type: object
    payment_SubmitOrderRequest:
      description: This object is expected as input
      properties:
        userId:
          description: generated token used to identify the caller of the endpoint
          type: string
        listOfItems:
          items:
            $ref: '#/components/schemas/ItemObject'
          type: array
        discount:
          type: number
        storeId:
          type: string
        orderType:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        deliveryAddress:
          type: string
      type: object
    payment_SubmitOrderResponse:
      description: This object is expected as output
      example:
        orderId: orderId
        success: true
        message: message
        timestamp: timestamp
      properties:
        orderId:
          type: string
        message:
          description: Will contain a descriptive error message if response code is
            not 200, else will be empty
          type: string
        success:
          type: boolean
        timestamp:
          type: string
      type: object
...
    payment_GetStatusRequest:
      description: This object is expected as input
      properties:
        orderID:
          type: string
      type: object
    payment_GetStatusResponse:
      description: This object is expected as output
      example:
        success: true
        message: message
        status: status
        timestamp: timestamp
      properties:
        status:
          type: string
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
      type: object
    payment_GetItemsRequest:
      description: This object is expected as input
      properties:
        orderID:
          type: string
      type: object
    payment_GetItemsResponse:
      description: This object is expected as output
      example:
        success: true
        message: message
        items:
        - itemType: itemType
          quantity: 6
          productId: productId
          size: size
          price: 0.8008281904610115
          imageUrl: imageUrl
          name: name
          description: description
          storeId: storeId
          barcode: barcode
          brand: brand
        - itemType: itemType
          quantity: 6
          productId: productId
          size: size
          price: 0.8008281904610115
          imageUrl: imageUrl
          name: name
          description: description
          storeId: storeId
          barcode: barcode
          brand: brand
        timestamp: timestamp
      properties:
        items:
          items:
            $ref: '#/components/schemas/ItemObject'
          type: array
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
      type: object
...